requirements.txt
Flask==3.0.0
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.5.0
gunicorn==21.2.0
openai==1.3.7
pydub==0.25.1
python-dotenv==1.0.0
Werkzeug==3.0.1